backend/requirements.txt
Django >= 4.1.3
djangorestframework >= 3.14.0
django-cors-headers >= 4.3.1
psycopg2 >= 2.9.9
pillow >= 10.2.0
psycopg2-binary >= 2.9.9
django-cors-headers >= 4.3.1